Update LuCI app for OpenWrt 25.12

+++ Makefile
@@ -1,5 +1,6 @@
 #
-# Copyright (C) 2016-2023 honwen https://github.com/honwen
+# Copyright (C) 2016-2026 honwen https://github.com/honwen
+# Copyright (C) 2026 woffko
 #
 # This is free software, licensed under the MIT License.
 # See /LICENSE for more information.
@@ -8,16 +9,18 @@
 include $(TOPDIR)/rules.mk
 
 PKG_NAME:=luci-app-haproxy-tcp
-PKG_VERSION:=0.4.0
-PKG_RELEASE:=3
+PKG_VERSION:=1.0.0
+PKG_RELEASE:=1
 
 PKG_LICENSE:=MIT
 PKG_LICENSE_FILES:=LICENSE
-PKG_MAINTAINER:=honwen <https://github.com/honwen>
+PKG_MAINTAINER:=woffko <https://github.com/woffko>
 
-TITLE:=LuCI Support for HAProxy-TCP
-LUCI_DEPENDS:=+haproxy-nossl
+LUCI_TITLE:=LuCI support for HAProxy TCP load balancing
+LUCI_DEPENDS:=+luci-base +haproxy
 LUCI_PKGARCH:=all
+LUCI_URL:=https://github.com/woffko/luci-app-haproxy-tcp
+LUCI_MAINTAINER:=woffko <https://github.com/woffko>
 
 define Package/$(PKG_NAME)/conffiles
 /etc/config/haproxy-tcp
@@ -27,12 +30,12 @@ include $(TOPDIR)/feeds/luci/luci.mk
 
 define Package/$(PKG_NAME)/postinst
 #!/bin/sh
-if [ -z "$${IPKG_INSTROOT}" ]; then
+if [ -z "$${IPKG_INSTROOT}" ] && [ -z "$${APK_INSTROOT}" ]; then
 	if [ -f /etc/uci-defaults/luci-haproxy-tcp ]; then
 		( . /etc/uci-defaults/luci-haproxy-tcp ) && \
 		rm -f /etc/uci-defaults/luci-haproxy-tcp
 	fi
-	rm -rf /tmp/luci-indexcache /tmp/luci-modulecache
+	rm -f /tmp/luci-indexcache /tmp/luci-modulecache
 fi
 exit 0
 endef
